[2018.12.18 김은비] 마이페이지 버그 수정

+++ marryus/src/main/java/kr/co/marryus/repository/domain/Auction.java
@@ -33,7 +33,7 @@ public class Auction {
 	
 	private int pageNo;
 	
-	private int dday;
+	private double dday;
 	
 	
 }
